# apps/DaisukiTTS-engine/test/e2e/__snapshots__/test_openapi/test_OpenAPIの形が変わっていないことを確認.json
{
  "components": {
    "schemas": {
...
      "AivmInfo": {
        "description": "AIVM (Aivis Voice Model) 仕様に準拠した音声合成モデルのメタデータ情報。\n\nAIVM マニフェストには、音声合成モデルに関連する全てのメタデータが含まれる。\nspeakers フィールド内の話者情報は、VOICEVOX ENGINE との API 互換性のために、\nAIVM マニフェストを基に Speaker / SpeakerStyle / SpeakerInfo / StyleInfo モデルに変換したもの。",
        "properties": {
          "file_path": {
            "format": "path",
            "title": "AIVMX ファイルのインストール先パス",
            "type": "string"
          },
          "file_size": {
            "title": "AIVMX ファイルのインストールサイズ (バイト単位)",
            "type": "integer"
          },
          "is_loaded": {
            "title": "この音声合成モデルがロードされているかどうか",
            "type": "boolean"
          },
          "is_private_model": {
            "title": "AivisHub で公開されておらず、ユーザーがローカルからインストールしたモデルの場合は True (ネットワークエラーなどで AivisHub から情報を取得できなかった場合も True を返す)",
            "type": "boolean"
          },
          "is_update_available": {
            "title": "この音声合成モデルの新しいバージョンが AivisHub で公開されているかどうか",
            "type": "boolean"
          },
          "latest_version": {
            "title": "この音声合成モデルの AivisHub で公開されている最新バージョン (AivisHub で公開されていない場合は AIVM マニフェスト記載のバージョン)",
            "type": "string"
          },
          "manifest": {
            "$ref": "#/components/schemas/AivmManifest",
            "title": "AIVM マニフェスト"
          },
          "speakers": {
            "items": {
              "$ref": "#/components/schemas/LibrarySpeaker"
            },
            "title": "話者情報のリスト (VOICEVOX ENGINE 互換)",
            "type": "array"
          }
        },
        "required": [
          "is_loaded",
          "is_update_available",
          "is_private_model",
          "latest_version",
          "file_path",
          "file_size",
          "manifest",
          "speakers"
        ],
        "title": "AivmInfo",
        "type": "object"
      },
      "AivmManifest": {
        "description": "AIVM マニフェストのスキーマ",
        "properties": {
          "creators": {
            "default": [],
            "items": {
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "title": "Creators",
            "type": "array"
          },
          "description": {
            "default": "",
            "maxLength": 140,
            "title": "Description",
            "type": "string"
          },
          "license": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "manifest_version": {
            "const": "1.0",
            "title": "Manifest Version",
            "type": "string"
          },
          "model_architecture": {
            "$ref": "#/components/schemas/ModelArchitecture"
          },
          "model_format": {
            "$ref": "#/components/schemas/ModelFormat"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "speakers": {
            "items": {
              "$ref": "#/components/schemas/AivmManifestSpeaker"
            },
            "title": "Speakers",
            "type": "array"
          },
          "training_epochs": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Epochs"
          },
          "training_steps": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Steps"
          },
          "uuid": {
            "format": "uuid",
            "title": "Uuid",
            "type": "string"
          },
          "version": {
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-((?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\\.(?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\\+([0-9a-zA-Z-]+(?:\\.[0-9a-zA-Z-]+)*))?$",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "manifest_version",
          "name",
          "model_architecture",
          "model_format",
          "uuid",
          "version",
          "speakers"
        ],
        "title": "AivmManifest",
        "type": "object"
      },
...
      "Body_install_model_aivm_models_install_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "format": "binary",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "AIVMX ファイル (`.aivmx`)",
            "title": "File"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "AIVMX ファイルの URL",
            "title": "Url"
          }
        },
        "title": "Body_install_model_aivm_models_install_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LibrarySpeaker": {
        "description": "音声ライブラリに含まれるキャラクターの情報。",
        "properties": {
          "speaker": {
            "$ref": "#/components/schemas/Speaker"
          },
          "speaker_info": {
            "$ref": "#/components/schemas/SpeakerInfo"
          }
        },
        "required": [
          "speaker",
          "speaker_info"
        ],
        "title": "LibrarySpeaker",
        "type": "object"
      },
...
      "ModelArchitecture": {
        "enum": [
          "Style-Bert-VITS2",
          "Style-Bert-VITS2 (JP-Extra)"
        ],
        "title": "ModelArchitecture",
        "type": "string"
      },
      "ModelFormat": {
        "enum": [
          "Safetensors",
          "ONNX"
        ],
        "title": "ModelFormat",
        "type": "string"
      },
...
      "Speaker": {
        "description": "キャラクター情報",
        "properties": {
          "name": {
            "description": "名前",
            "title": "Name",
            "type": "string"
          },
          "speaker_uuid": {
            "description": "キャラクターの UUID",
            "title": "Speaker Uuid",
            "type": "string"
          },
          "styles": {
            "description": "スタイルの一覧",
            "items": {
              "$ref": "#/components/schemas/SpeakerStyle"
            },
            "title": "Styles",
            "type": "array"
          },
          "supported_features": {
            "$ref": "#/components/schemas/SpeakerSupportedFeatures",
            "description": "キャラクターの対応機能"
          },
          "version": {
            "description": "キャラクターのバージョン",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "speaker_uuid",
          "styles",
          "version"
        ],
        "title": "Speaker",
        "type": "object"
      },
      "SpeakerInfo": {
        "description": "キャラクターの追加情報",
        "properties": {
          "policy": {
            "description": "policy.md",
            "title": "Policy",
            "type": "string"
          },
          "portrait": {
            "description": "アイコンを base64 エンコードしたもの、あるいは URL (「立ち絵画像を base64 エンコードしたもの」ではない点で VOICEVOX ENGINE と異なる)",
            "title": "Portrait",
            "type": "string"
          },
          "style_infos": {
            "description": "スタイルの追加情報",
            "items": {
              "$ref": "#/components/schemas/StyleInfo"
            },
            "title": "Style Infos",
            "type": "array"
          }
        },
        "required": [
          "policy",
          "portrait",
          "style_infos"
        ],
        "title": "SpeakerInfo",
        "type": "object"
      },
      "SpeakerStyle": {
        "description": "キャラクターのスタイル情報",
        "properties": {
          "id": {
            "description": "スタイル ID",
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "description": "スタイル名",
            "title": "Name",
            "type": "string"
          },
          "type": {
            "default": "talk",
            "description": "スタイルの種類。talk: 音声合成クエリの作成と音声合成が可能。singing_teacher: 歌唱音声合成用のクエリの作成が可能。frame_decode: 歌唱音声合成が可能。sing: 歌唱音声合成用のクエリの作成と歌唱音声合成が可能。",
            "enum": [
              "talk",
              "singing_teacher",
              "frame_decode",
              "sing"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "id"
        ],
        "title": "SpeakerStyle",
        "type": "object"
      },
      "SpeakerSupportedFeatures": {
        "description": "キャラクターの対応機能の情報",
        "properties": {
          "permitted_synthesis_morphing": {
            "default": "ALL",
            "description": "モーフィング機能への対応。'ALL' は「全て許可」、'SELF_ONLY' は「同じキャラクター内でのみ許可」、'NOTHING' は「全て禁止」",
            "enum": [
              "ALL",
              "SELF_ONLY",
              "NOTHING"
            ],
            "title": "Permitted Synthesis Morphing",
            "type": "string"
          }
        },
        "title": "SpeakerSupportedFeatures",
        "type": "object"
      },
      "StyleInfo": {
        "description": "スタイルの追加情報",
        "properties": {
          "icon": {
            "description": "このスタイルのアイコンを base64 エンコードしたもの、あるいは URL",
            "title": "Icon",
            "type": "string"
          },
          "id": {
            "description": "スタイル ID",
            "title": "Id",
            "type": "integer"
          },
          "portrait": {
            "description": "AivisSpeech Engine では常に None を返す (「このスタイルの立ち絵画像を base64 エンコードしたもの」ではない点で VOICEVOX ENGINE と異なる)",
            "title": "Portrait",
            "type": "string"
          },
          "voice_sample_transcripts": {
            "default": [],
            "description": "ボイスサンプルの書き起こしテキスト (voice_samples の配列インデックスと対応し、存在しない場合は空文字列)",
            "items": {
              "type": "string"
            },
            "title": "Voice Sample Transcripts",
            "type": "array"
          },
          "voice_samples": {
            "description": "ボイスサンプルの音声データを base64 エンコードしたもの、あるいは URL",
            "items": {
              "type": "string"
            },
            "title": "Voice Samples",
            "type": "array"
          }
        },
        "required": [
          "id",
          "icon",
          "voice_samples"
        ],
        "title": "StyleInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AivisSpeech の音声合成エンジンです。",
    "title": "AivisSpeech Engine",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/aivm_models/install": {
      "post": {
        "description": "音声合成モデルをインストールします。<br>\nファイルからインストールする場合は `file` を指定してください。<br>\nURL からインストールする場合は `url` を指定してください。",
        "operationId": "install_model",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_install_model_aivm_models_install_post"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "音声合成モデルをインストールする",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
    "/aivm_models/{aivm_uuid}": {
      "get": {
        "description": "指定された音声合成モデルの情報を取得します。",
        "operationId": "get_aivm_info",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AivmInfo"
                }
              }
            },
            "description": "指定された音声合成モデルの情報"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定された音声合成モデルの情報を取得する",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
    "/aivm_models/{aivm_uuid}/load": {
      "post": {
        "description": "指定された音声合成モデルをロードします。すでにロード済みの場合は何も行われません。<br>\n実行しなくても他の API は利用できますが、音声合成の初回実行時に時間がかかることがあります。",
        "operationId": "load_model",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定された音声合成モデルをロードする",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
    "/aivm_models/{aivm_uuid}/uninstall": {
      "delete": {
        "description": "指定された音声合成モデルをアンインストールします。",
        "operationId": "uninstall_model",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定された音声合成モデルをアンインストールする",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
    "/aivm_models/{aivm_uuid}/unload": {
      "post": {
        "description": "指定された音声合成モデルをアンロードします。",
        "operationId": "unload_model",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定された音声合成モデルをアンロードする",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
    "/aivm_models/{aivm_uuid}/update": {
      "post": {
        "description": "AivisHub から指定された音声合成モデルの一番新しいバージョンをダウンロードし、\nインストール済みの音声合成モデルへ上書き更新します。",
        "operationId": "update_model",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定された音声合成モデルを更新する",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
...
    "/singer_info": {
      "get": {
        "operationId": "singer_info",
        "parameters": [
          {
            "in": "query",
            "name": "speaker_uuid",
            "required": true,
            "schema": {
              "title": "Speaker Uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resource_format",
            "required": false,
            "schema": {
              "default": "base64",
              "enum": [
                "base64",
                "url"
              ],
              "title": "Resource Format",
              "type": "string"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeakerInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "AivisSpeech Engine ではサポートされていない API です (常に 501 Not Implemented を返します)",
        "tags": [
          "話者情報"
        ]
      }
    },
    "/singers": {
      "get": {
        "operationId": "singers",
        "parameters": [
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Speaker"
                  },
                  "title": "Response Singers Singers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "AivisSpeech Engine ではサポートされていない API です (常に 501 Not Implemented を返します)",
        "tags": [
          "話者情報"
        ]
      }
    },
    "/speaker_info": {
      "get": {
        "description": "UUID で指定された話者の情報を返します。\n画像や音声は resource_format で指定した形式で返されます。",
        "operationId": "speaker_info",
        "parameters": [
          {
            "in": "query",
            "name": "speaker_uuid",
            "required": true,
            "schema": {
              "title": "Speaker Uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resource_format",
            "required": false,
            "schema": {
              "default": "base64",
              "enum": [
                "base64",
                "url"
              ],
              "title": "Resource Format",
              "type": "string"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeakerInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "UUID で指定された話者の情報を取得する",
        "tags": [
          "話者情報"
        ]
      }
    },
    "/speakers": {
      "get": {
        "description": "話者情報の一覧を返します。",
        "operationId": "speakers",
        "parameters": [
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Speaker"
                  },
                  "title": "Response Speakers Speakers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "話者情報の一覧を取得する",
        "tags": [
          "話者情報"
        ]
      }
    },
...
  }
}